+++ app/Http/Controllers/api/ProductController.php
@@ -17,21 +17,25 @@ class ProductController extends Controller
         try {
             $products = Product::all();
             if ($products->isEmpty()) {
-                return APIFormatter::createAPI(404, 'error', 'Products not found', null);
+                return APIFormatter::createAPI(200, 'success', 'Products not found', null);
             }
             return APIFormatter::createAPI(200, 'success', 'Products found', $products);
         } catch (\Throwable $e) {
-            return APIFormatter::createAPI(400, 'fail', 'Failed to get products', $e->getMessage());
+            return APIFormatter::createAPI(500, 'error', 'Failed to get products', $e->getMessage());
         }
     }
 
     public function show($id)
     {
-        $product = Product::find($id);
-        if (!$product) {
-            return APIFormatter::createAPI(404, 'error', 'Product not found', null);
+        try {
+            $product = Product::findOrFail($id);
+            if (!$product) {
+                return APIFormatter::createAPI(200, 'success', 'Product not found', null);
+            }
+            return APIFormatter::createAPI(200, 'success', 'Product found', $product);
+        } catch (\Throwable $th) {
+            return APIFormatter::createAPI(400, 'error', 'Failed to get product', $th->getMessage());
         }
-        return APIFormatter::createAPI(200, 'success', 'Product found', $product);
     }
 
     public function store(Request $request)
@@ -45,7 +49,7 @@ class ProductController extends Controller
                 'minimum_order' => 'required|integer',
                 'description' => 'nullable|string',
                 'product_category_id' => 'required|exists:product_categories,id',
-                // 'merchant_id' => 'required|exists:merchants,id'
+                'merchant_id' => 'required|exists:merchants,id'
             ]);
             $product = Product::create([
                 'name' => $request->name,
@@ -54,7 +58,7 @@ class ProductController extends Controller
                 'minimum_order' => $request->minimum_order,
                 'description' => $request->description,
                 'product_category_id' => $request->product_category_id,
-                // 'merchant_id' => $request->merchant_id
+                'merchant_id' => $request->merchant_id
             ]);
             DB::commit();
             return APIFormatter::createAPI(201, 'success', 'Product created', $product);
@@ -73,25 +77,29 @@ class ProductController extends Controller
             'minimum_order' => 'required|integer',
             'description' => 'nullable|string',
             'product_category_id' => 'required|exists:product_categories,id',
-            // 'merchant_id' => 'required|exists:merchants,id'
+            'merchant_id' => 'required|exists:merchants,id'
         ]);
         DB::beginTransaction();
         try {
             $product =  product::findOrFail($id);
+            if (!$product) {
+                return APIFormatter::createAPI(200, 'success', 'Product not found', null);
+            }
             $product->update([
                 'name' => $request->name,
                 'price' => $request->price,
                 'stock' => $request->stock,
                 'minimum_order' => $request->minimum_order,
                 'description' => $request->description,
-                'product_category_id' => $request->product_category_id
+                'product_category_id' => $request->product_category_id,
+                'merchant_id' => $request->merchant_id
             ]);
             DB::commit();
-        } catch (\Exception $e) {
+            return APIFormatter::createAPI(200, 'success', 'Product updated', $product);
+        } catch (\Throwable $e) {
             DB::rollBack();
-            return APIFormatter::createAPI(400, 'fail', 'Failed to update product', null);
+            return APIFormatter::createAPI(400, 'fail', 'Failed to update product', $e->getMessage());
         }
-        return APIFormatter::createAPI(200, 'success', 'Product updated', $product);
     }
 
     public function destroy($id)
@@ -99,13 +107,16 @@ class ProductController extends Controller
         DB::beginTransaction();
         try {
             $product = Product::findOrFail($id);
+            if (!$product) {
+                return APIFormatter::createAPI(200, 'success', 'Product not found', null);
+            }
             Product_Image::where('product_id', $id)->delete();
             $product->delete();
             DB::commit();
             return APIFormatter::createAPI(200, 'success', 'Product deleted', null);
-        } catch (\Exception $e) {
+        } catch (\Throwable $e) {
             DB::rollBack();
-            return APIFormatter::createAPI(400, 'fail', 'Failed to delete product', null);
+            return APIFormatter::createAPI(400, 'fail', 'Failed to delete product', $e->getMessage());
         }
     }
 }
